splitting matching service additional name spec into two files
